+++ src/main/java/com/tmall/wireless/tac/biz/processor/extremeItem/domain/ItemGmvGroupMap.java
@@ -6,9 +6,9 @@ import java.util.Map;
 
 @Data
 public class ItemGmvGroupMap {
-    private Map<Long, ItemGmvGroup> itemGmvGroupMap;
+    private Map<Integer, ItemGmvGroup> itemGmvGroupMap;
 
-    public double raceValueOf(Long groupNo) {
+    public double raceValueOf(Integer groupNo) {
         return itemGmvGroupMap.get(groupNo).ranceValue();
     }
